chore: rename self_hosted file
