+++ swanlab/api/self_hosted.py
