+++ swanlab/cli/api/self_hosted.py
